+++ src/worlds/map.rs
@@ -2,177 +2,148 @@ use std::collections::HashMap;
 use std::path::PathBuf;
 
 use bevy::prelude::*;
+use bevy::render::settings;
+use bevy::render::view::visibility;
+use bevy::sprite::{MaterialMesh2dBundle, Mesh2dHandle};
 use bevy::{math::uvec2, transform::commands};
-use bevy_fast_tilemap::{bundle::MapBundleManaged, map::Map, plugin::FastTileMapPlugin};
+use bevy_ecs_tilemap::prelude::*;
 use biomes::Biome;
 use events::{DrawMapEvent, GenerateMapEvent};
-use generator::elevation::ElevationGenerator;
-use generator::temperature::TemperatureGenerator;
+use generator::elevation::{generate as elevation_gen, ElevationGenerator, TileElevation};
+use generator::temperature::generate as temperature_gen;
+use generator::temperature::{TemperatureGenerator, TileTemperature};
 use generator::MapGenerator;
 use noise::{NoiseFn, Perlin};
 use renderer::elevation::ElevationMapRenderer;
 use renderer::temperature::TemperatureMapRenderer;
-use renderer::MapRenderer;
 use shapes::{CircleCenteredShape, ContinentsShape, ShapeGenerator, ShapeGeneratorResource};
-use tile::{Tile, TileMatrixResource};
+
+use super::settings::{MapMode, Settings};
 pub(crate) mod biomes;
 mod events;
 mod generator;
 mod renderer;
 mod shapes;
-mod tile;
-mod tileset;
-
-use super::{
-    settings::{MapMode, Settings, WorldShapeEnum},
-    utils::{scale_to_index, xy_to_lonlat},
-};
-use tileset::{TextureTile, TextureTileSet};
 
 const MAX_PERLIN_SCALE: f64 = 100000.;
 
-#[derive(Resource)]
-pub struct MapGeneratorsResource {
-    generators: Vec<Box<dyn MapGenerator>>,
+#[derive(Component)]
+struct LastUpdate(f64);
+
+#[derive(SystemSet, Debug, Clone, PartialEq, Eq, Hash)]
+enum MapSet {
+    Prepare,
+    Generate,
+    Render,
 }
 
-impl FromWorld for MapGeneratorsResource {
-    fn from_world(world: &mut World) -> Self {
-        let config = world.resource::<Settings>();
-
-        let mut generators: Vec<Box<dyn MapGenerator>> =
-            vec![Box::new(ElevationGenerator), Box::new(TemperatureGenerator)];
-
-        Self { generators }
-    }
-}
-
-#[derive(Resource)]
-struct MapRendererResource {
-    renderer: Box<dyn MapRenderer>,
-}
-
-impl FromWorld for MapRendererResource {
-    fn from_world(world: &mut World) -> Self {
-        let config = world.resource::<Settings>();
-        let texture_tileset = world.resource::<TextureTileSet>();
-
-        let renderer: Box<dyn MapRenderer> = match config.mode {
-            MapMode::Elevation => Box::new(ElevationMapRenderer::new(texture_tileset)),
-            MapMode::Temperature => Box::new(TemperatureMapRenderer::new(texture_tileset)),
-            MapMode::WorldShapeMode => Box::new(ElevationMapRenderer::new(texture_tileset)),
-        };
-
-        Self { renderer }
-    }
+#[derive(Debug, Clone, Copy, Default, Eq, PartialEq, Hash, States)]
+enum MapState {
+    #[default]
+    Normal,
 }
 
 pub(super) fn plugin(app: &mut App) {
-    app.init_resource::<TextureTileSet>()
-        .init_resource::<TileMatrixResource>()
-        .init_resource::<MapGeneratorsResource>()
-        .init_resource::<MapRendererResource>()
-        .add_plugins(FastTileMapPlugin::default())
-        .add_plugins((biomes::plugin, shapes::plugin))
+    app.add_plugins(TilemapPlugin)
+        .init_state::<MapState>()
+        .configure_sets(
+            Update,
+            (MapSet::Prepare, MapSet::Generate, MapSet::Render).chain(),
+        )
         .add_systems(Startup, setup_map)
-        .add_systems(Update, update_map)
-        .observe(on_generate_map)
-        .observe(on_draw_map);
+        .add_systems(
+            Update,
+            (
+                (elevation_gen, temperature_gen).in_set(MapSet::Generate),
+                (update_tiles_color).in_set(MapSet::Render),
+            )
+                .run_if(resource_changed::<Settings>),
+        );
 }
 
-fn setup_map(
-    mut commands: Commands,
-    config: Res<Settings>,
-    texture_tileset: Res<TextureTileSet>,
-    asset_server: Res<AssetServer>,
-    mut materials: ResMut<Assets<Map>>,
-    maps: Query<&Handle<Map>>,
-    mut shape_generator_resource: ResMut<ShapeGeneratorResource>,
-) {
-    let tiles_texture = asset_server.load(texture_tileset.path.clone());
+fn setup_map(mut commands: Commands, config: Res<Settings>, asset_server: Res<AssetServer>) {
+    let map_size = TilemapSize {
+        x: config.width,
+        y: config.height,
+    };
 
-    let map = Map::builder(
-        // Map size
-        uvec2(config.width, config.height),
-        // Tile atlas
-        tiles_texture,
-        // Tile Size
-        config.tile_size,
-    )
-    .build_and_set(|_| 2);
+    let tilemap_entity = commands.spawn_empty().id();
+    let texture_handle: Handle<Image> = asset_server.load("tiles_white.png");
+    let coord_sys: HexCoordSystem = HexCoordSystem::RowEven;
 
-    commands.spawn(MapBundleManaged {
-        material: materials.add(map),
-        ..default()
-    });
-
-    commands.trigger(GenerateMapEvent);
-}
-
-fn on_generate_map(
-    trigger: Trigger<GenerateMapEvent>,
-    mut commands: Commands,
-    mut tile_matrix: ResMut<TileMatrixResource>,
-    config: Res<Settings>,
-    mut generators: ResMut<MapGeneratorsResource>,
-) {
-    for x in 0..config.width {
-        for y in 0..config.height {
-            let mut tile: Tile = Tile::default();
-
-            for generator in &generators.generators {
-                generator.apply(&mut tile, x, y, &config);
-            }
-
-            tile_matrix.set(x as usize, y as usize, tile);
+    // Initialize tile storage with empty tiles
+    let mut tile_storage = TileStorage::empty(map_size);
+    for x in 0..map_size.x {
+        for y in 0..map_size.y {
+            let tile_pos = TilePos { x, y };
+            let tile_entity = commands
+                .spawn(TileBundle {
+                    position: tile_pos,
+                    tilemap_id: TilemapId(tilemap_entity),
+                    color: TileColor(Color::srgb(0., 0., 0.)),
+                    ..default()
+                })
+                .id();
+            tile_storage.set(&tile_pos, tile_entity);
         }
     }
 
-    println!("{:?}", tile_matrix);
+    let tile_size = TilemapTileSize {
+        x: config.tile_size.x,
+        y: config.tile_size.y,
+    };
 
-    commands.trigger(DrawMapEvent);
+    let grid_size = tile_size.into();
+    let map_type = TilemapType::Hexagon(coord_sys);
+
+    commands.entity(tilemap_entity).insert((
+        TilemapBundle {
+            grid_size,
+            map_type,
+            tile_size,
+            size: map_size,
+            storage: tile_storage,
+            texture: TilemapTexture::Single(texture_handle.clone()),
+            transform: get_tilemap_center_transform(&map_size, &grid_size, &map_type, 0.0),
+            ..default()
+        },
+        LastUpdate(-1.0),
+    ));
 }
 
-fn on_draw_map(
-    trigger: Trigger<DrawMapEvent>,
-    asset_server: Res<AssetServer>,
-    texture_tileset: Res<TextureTileSet>,
-    mut materials: ResMut<Assets<Map>>,
-    config: Res<Settings>,
-    maps: Query<&Handle<Map>>,
-    tile_matrix: Res<TileMatrixResource>,
-    renderer: Res<MapRendererResource>,
+fn update_tiles_color(
+    time: Res<Time>,
+    settings: Res<Settings>,
+    mut tilemap_query: Query<(&mut Visibility, &mut LastUpdate)>,
+    mut tile_query: Query<(&TileElevation, &TileTemperature, &mut TileColor)>,
 ) {
-    for map_handle in maps.iter() {
-        let map = materials.get_mut(map_handle).unwrap();
-        let mut m = map.indexer_mut();
+    let mut elevation_renderer = ElevationMapRenderer;
+    let mut temperature_renderer = TemperatureMapRenderer;
+    let current_time = time.elapsed_seconds_f64();
+    for (mut visibility, mut last_update) in tilemap_query.iter_mut() {
+        if current_time - last_update.0 > 0.1 {
+            if !settings.elevation && !settings.temperature {
+                *visibility = Visibility::Hidden;
+            } else {
+                *visibility = Visibility::Visible;
 
-        for x in 0..tile_matrix.width {
-            for y in 0..tile_matrix.height {
-                let tile = tile_matrix.get(x, y).unwrap();
-                let tile_index = renderer.renderer.get_tile_index(tile);
+                for (tile_elevation, tile_temperature, mut tile_color) in tile_query.iter_mut() {
+                    let mut color = Color::srgba(1., 1., 1., 1.);
 
-                m.set(x.try_into().unwrap(), y.try_into().unwrap(), tile_index);
+                    if settings.elevation {
+                        color.mix_assign(elevation_renderer.get_color(tile_elevation), 1.);
+                    }
+                    if settings.temperature {
+                        color.mix_assign(
+                            temperature_renderer.get_color(tile_temperature, &settings),
+                            settings.temperature_factor,
+                        );
+                    }
+
+                    *tile_color = TileColor(color);
+                }
             }
+            last_update.0 = current_time;
         }
-    }
-}
-
-fn update_map(
-    mut commands: Commands,
-    config: Res<Settings>,
-    mut materials: ResMut<Assets<Map>>,
-    maps: Query<&Handle<Map>>,
-    mut shape_generator_resource: ResMut<ShapeGeneratorResource>,
-    mut renderer: ResMut<MapRendererResource>,
-    texture_tileset: Res<TextureTileSet>,
-) {
-    if config.is_changed() {
-        renderer.renderer = match config.mode {
-            MapMode::Elevation => Box::new(ElevationMapRenderer::new(&texture_tileset)),
-            MapMode::Temperature => Box::new(TemperatureMapRenderer::new(&texture_tileset)),
-            MapMode::WorldShapeMode => Box::new(ElevationMapRenderer::new(&texture_tileset)),
-        };
-        commands.trigger(GenerateMapEvent);
     }
 }